requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
et-xmlfile==1.1.0
fastapi==0.110.1
fastapi-pagination==0.12.23
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pypika-tortoise==0.1.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
tomlkit==0.12.4
tortoise==0.1.1
tortoise-orm==0.20.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
